requirements.txt
annotated-types==0.7.0
anyio==4.11.0
backports-zstd==1.0.0
blinker==1.9.0
brotli==1.2.0
cachelib==0.13.0
certifi==2025.10.5
cffi==2.0.0
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
deprecation==2.1.0
dotenv==0.9.9
flask==3.1.2
flask-compress==1.23
flask-minify==0.50
flask-session==0.8.0
flask-talisman==1.1.0
flask-wtf==1.2.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
htmlminf==0.1.13
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
jinja2==3.1.6
jsmin==3.0.1
lesscpy==0.15.1
markupsafe==3.0.3
msgspec==0.19.0
multidict==6.7.0
nh3==0.3.2
packaging==25.0
pillow==12.0.0
ply==3.11
postgrest==2.24.0
propcache==0.4.1
pycparser==2.23
pydantic==2.12.4
pydantic-core==2.41.5
pyjwt==2.10.1
python-dotenv==1.2.1
rcssmin==1.2.2
realtime==2.24.0
six==1.17.0
sniffio==1.3.1
storage3==2.24.0
strenum==0.4.15
supabase==2.24.0
supabase-auth==2.24.0
supabase-functions==2.24.0
typing-extensions==4.15.0
typing-inspection==0.4.2
websockets==15.0.1
werkzeug==3.1.3
wtforms==3.2.1
xxhash==3.6.0
yarl==1.22.0
